Changed activities to fragments, added exit menu
