+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/MyTheme"
         tools:targetApi="32">
-        <activity
-            android:name=".DetailsActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
+
         <activity
             android:name=".MainActivity"
             android:exported="true">
