+++ app/src/main/java/com/bolunevdev/kinon/TopSpacingItemDecoration.kt
@@ -6,11 +6,16 @@ import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 
 
-class TopSpacingItemDecoration (private val paddingInDp: Int): RecyclerView.ItemDecoration() {
+class TopSpacingItemDecoration(private val paddingInDp: Int) : RecyclerView.ItemDecoration() {
     private val Int.convertPx: Int
         get() = (this * Resources.getSystem().displayMetrics.density).toInt()
 
-    override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
+    override fun getItemOffsets(
+        outRect: Rect,
+        view: View,
+        parent: RecyclerView,
+        state: RecyclerView.State
+    ) {
         super.getItemOffsets(outRect, view, parent, state)
         outRect.top = paddingInDp.convertPx
         outRect.right = paddingInDp.convertPx
