+++ app/src/main/res/layout/fragment_exit_menu.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    style="@style/activity_main_background"
+    android:orientation="vertical">
+    <TextView
+        android:id="@+id/text_header"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:layout_margin="@dimen/exit_menu_text_header_margin"
+        android:textColor="@android:color/white"
+        android:textSize="@dimen/exit_menu_text_header_textSize"
+        android:text="@string/exit_menu_text_header">
+    </TextView>
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+        <Button
+            android:id="@+id/btn_y"
+            android:layout_width="0dp"
+            android:layout_weight="50"
+            android:textColor="@color/white"
+            android:layout_height="wrap_content"
+            android:layout_margin="5dp"
+            android:text="@string/exit_menu_btn_y"
+            style="?android:attr/buttonBarButtonStyle">
+        </Button>
+        <Button
+            android:id="@+id/btn_n"
+            android:layout_width="0dp"
+            android:layout_weight="50"
+            android:textColor="@color/white"
+            android:layout_height="wrap_content"
+            android:layout_margin="5dp"
+            android:text="@string/exit_menu_btn_n"
+            style="?android:attr/buttonBarButtonStyle">
+        </Button>
+    </LinearLayout>
+</LinearLayout>
